--- titus-supplementary-component/job-activity-history/src/main/java/com/netflix/titus/supplementary/jobactivity/store/JobActivityConsumerStore.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2019 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.netflix.titus.supplementary.jobactivity.store;
-
-import com.netflix.titus.api.common.LeaderActivationListener;
-import com.netflix.titus.api.jobmanager.model.job.Job;
-import com.netflix.titus.api.jobmanager.model.job.Task;
-
-/*
-The consumer store consumes from the publisher queue
- */
-public interface JobActivityConsumerStore extends LeaderActivationListener {
-    public void consumeJob(Job<?> Job);
-    public void consumeTask(Task Task);
-    public void consumeRecord();
-}
