+++ titus-supplementary-component/job-activity-history/src/main/java/com/netflix/titus/supplementary/jobactivity/store/JobActivityStore.java
@@ -16,8 +16,16 @@
 
 package com.netflix.titus.supplementary.jobactivity.store;
 
-import com.netflix.titus.common.runtime.TitusRuntime;
+import com.netflix.titus.api.common.LeaderActivationListener;
+import com.netflix.titus.api.jobmanager.model.job.Job;
+import com.netflix.titus.api.jobmanager.model.job.Task;
 
-public interface JobActivityStore {
+/*
+The consumer store consumes from the publisher queue
+ */
 
+public interface JobActivityStore extends LeaderActivationListener {
+    public void consumeJob(Job<?> Job);
+    public void consumeTask(Task Task);
+    public void consumeRecord();
 }
